requirements.txt
aioredis==1.3.1
asgiref==2.3.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.1
channels==2.0.0
channels-redis==2.2.1
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==1.11.17
django-cors-headers==3.0.0
django-dynamic-fixture==3.1.0
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.5.6
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
wrapt==1.12.1
zope.interface==5.1.0
